+++ tests/testthat/test-cql-string.R
@@ -63,6 +63,10 @@ test_that("All cql geom predicate functions work", {
     BBOX(c(1,2,1,2), crs = 'EPSG:4326'),
     CQL("BBOX({geom_name}, 1, 2, 1, 2, 'EPSG:4326')")
   )
+  expect_equal(
+    BBOX(c(1,2,1,2), crs = 4326),
+    CQL("BBOX({geom_name}, 1, 2, 1, 2, 'EPSG:4326')")
+  )
 })
 
 test_that("CQL functions fail correctly", {
@@ -75,8 +79,6 @@ test_that("CQL functions fail correctly", {
   expect_error(RELATE(the_geom, rep("TTTTTTTTT", 2)), "pattern") # > length 1
   expect_error(BBOX(c(1,2,3)), "numeric vector")
   expect_error(BBOX(c("1","2","3", "4")), "numeric vector")
-  expect_error(BBOX(c(1,2,3,4), crs = 4326), "must be a character string")
-  expect_error(BBOX(c(1,2,3,4), crs = 4326), "must be a character string")
   expect_error(BBOX(c(1,2,3,4), crs = c("EPSG:4326", "EPSG:3005")),
                "must be a character string")
 })