+++ NEWS.md
@@ -1,5 +1,7 @@
 # bcdata (development version)
 
+* Geometry predicates can now take a `bbox` object as well as an `sf*` object (#176)
+
 # bcdata 0.1.2
 
 ### IMPROVEMENTS